Move poll_until into _retry
