+++ flocker/common/__init__.py
@@ -8,7 +8,7 @@ __all__ = [
     'INode', 'FakeNode', 'ProcessNode', 'gather_deferreds',
     'auto_threaded', 'interface_decorator',
     'get_all_ips', 'ipaddress_from_string',
-    'loop_until', 'retry_failure',
+    'loop_until', 'retry_failure', 'poll_until',
 ]
 
 from ._ipc import INode, FakeNode, ProcessNode
@@ -16,4 +16,4 @@ from ._defer import gather_deferreds
 from ._thread import auto_threaded
 from ._interface import interface_decorator
 from ._net import get_all_ips, ipaddress_from_string
-from ._retry import loop_until, retry_failure
+from ._retry import loop_until, poll_until, retry_failure
